deployment/requirements.txt
gradio==3.50.2
torch==2.1.0
torchvision==0.16.0
Pillow==10.1.0
numpy==1.24.3
opencv-python-headless==4.8.1.78
